reformatted output functions into multiple modules
