+++ support/output_formatters.py
@@ -7,160 +7,6 @@ outputted to a text file.
 Author: Rani Hinnawi
 Date: 2023-08-08
 """
-from sys import setrecursionlimit
-from typing import Optional, List, Tuple
-from support.performance import Performance
-from lab3.huffman_tree import HuffmanTree
-from lab3.huffman_node import HuffmanNode
-
-# Set recursion limit. Python default: 1000. Should not need to exceed number
-# of letters in Latin (English) alphabet
-RECURSION_LIMIT = 52
-setrecursionlimit(RECURSION_LIMIT)
-
-
-def format_huffman_tree(huffman_tree: 'HuffmanTree', nodes_per_line: int) \
-        -> str:
-    """
-    Function that formats the number of Huffman Tree nodes per line for easier
-    readability in an output file.
-
-    Args:
-        huffman_tree (HuffmanTree): Huffman Tree being printed out
-        nodes_per_line (int): max number of Huffman Nodes printed per line
-
-    Returns:
-        str: string representation of Huffman Nodes
-
-    Raises:
-        ValueError: when nodes_per_line is not a positive integer
-    """
-    def preorder(root: Optional['HuffmanNode'], nodes_left=nodes_per_line) \
-            -> Tuple[List[str], int]:
-        """
-        Preorder traversal of Huffman Tree and its nodes.
-
-        Args:
-            root (HuffmanNode): root of a Huffman tree or subtree OR None
-
-        Returns:
-            List[str]: list of Python nodes as key-value pairs in format
-                character: frequency
-        """
-        # Add nodes to 'representation' list
-        rep = []
-
-        # Visit root, then left, then right
-        if not root:
-            return rep, nodes_left
-
-        node_str = str(root)
-
-        # Node is placed on new line when 0 nodes left per line are remaining
-        if nodes_left == 0:
-            node_str = "\n\t" + node_str
-            nodes_left = nodes_per_line
-        else:
-            nodes_left -= 1
-
-        # Add HuffmanNode string, then visit left and right branches
-        rep.append(node_str)
-
-        if root.get_left():
-            left, nodes_left = preorder(root.get_left(), nodes_left)
-            rep.extend(left)
-
-        if root.get_right():
-            right, nodes_left = preorder(
-                root.get_right(), nodes_left)
-            rep.extend(right)
-
-        # Returns preorder representation list and remaining nodes for
-        # possible recursive calls
-        return rep, nodes_left
-
-    if nodes_per_line < 1:
-        error = "There must be at least 1 node per line"
-        raise ValueError(error)
-
-    output, _ = preorder(huffman_tree.get_root())
-    return ", ".join(output)
-
-
-def format_huffman_tree_binary_codes(huffman_tree: 'HuffmanTree',
-                                     nodes_per_line: int) -> str:
-    """
-    Function that formats the number of Huffman Tree leaf nodes and their
-    respective binary Huffman Codes per line for easier readability in an
-    output file. 
-
-    Args:
-        huffman_tree (HuffmanTree): Huffman Tree codes being printed out
-        nodes_per_line (int): max number of Huffman Nodes printed per line
-
-    Returns:
-        str: string representation of Huffman Nodes
-
-    Raises:
-        ValueError: when nodes_per_line is not a positive integer
-    """
-    def preorder(root: Optional['HuffmanNode'], nodes_left=nodes_per_line) \
-            -> Tuple[List[str], int]:
-        """
-        Preorder traversal of Huffman Tree and its nodes. Only outputs leaf
-        nodes.
-
-        Args:
-            root (HuffmanNode): root of a Huffman tree or subtree OR None
-
-        Returns:
-            List[str]: list of Python nodes as key-value pairs in format
-                character: Huffman Code
-        """
-        # Add nodes to 'representation' list
-        rep = []
-
-        # Check root, then visit left and right subtrees
-        if not root:
-            # Case: Node (root) does not exist
-            return rep, nodes_left
-
-        if root.is_leaf():
-            # Case: root is a leaf. Can add it and its code to output
-            # node_str = f"{root.get_characters()}: {root.get_code()}"
-            node_str = f"{root} - {root.get_code()}"
-
-            # Node is placed on new line when 0 nodes left per line are remaining
-            if nodes_left == 0:
-                node_str = "\n\t" + node_str
-                nodes_left = nodes_per_line
-            else:
-                nodes_left -= 1
-
-            rep.append(node_str)
-            return rep, nodes_left
-        else:
-            # Get left subtree leaf nodes
-            if root.get_left():
-                left, nodes_left = preorder(root.get_left(), nodes_left)
-                rep.extend(left)
-
-            # Get right subtree leaf nodes
-            if root.get_right():
-                right, nodes_left = preorder(
-                    root.get_right(), nodes_left)
-                rep.extend(right)
-
-            # Returns preorder representation list and remaining nodes for
-            # possible recursive calls
-            return rep, nodes_left
-
-    if nodes_per_line < 1:
-        error = "There must be at least 1 node per line"
-        raise ValueError(error)
-
-    output, _ = preorder(huffman_tree.get_root())
-    return '\t' + ", ".join(output)
 
 
 def break_string(expression: str, chars_per_line: int) -> str:
@@ -250,54 +96,5 @@ def format_decoded_results(line_number: int, expression: str, result: str,
 
     # Add metrics and return
     write.append(f"{metrics}\n")
-
-    return '\n'.join(write)
-
-
-def format_performance_report(metrics: 'Performance', micro_sec=True) -> str:
-    """
-    Function that formats the size and runtime data logged for each success and
-    failure into a report. Runtimes are outputted by size in order from 
-    smallest to largest.
-
-    Args:
-        metrics (Performance): Performance object with logged metrics data
-        micro_sec (bool): True if saving runtime in microseconds, otherwise
-            False
-
-    Returns:
-        str: logged successes and failures, formatted to suit a text file
-
-    TODO: Update to have write be a List that is joined by \n
-    """
-    write = ["\n-------Performance Report-------\n"]
-
-    # Note total number of successes and successes per size
-    write.append(f"Total number of successes: {metrics.get_num_successes()}")
-
-    successes = metrics.get_successes()
-    for size in sorted(successes.keys()):
-        runtime = sorted(successes[size])
-        write.append(f"{size}: {runtime}")
-
-    write.append("\n")
-
-    # Note total number of errors and errors per size
-    write.append(f"Total number of errors: {metrics.get_num_errors()}")
-
-    errors = metrics.get_errors()
-    for size in sorted(errors.keys()):
-        runtime = sorted(errors[size])
-        write.append(f"{size}: {runtime}\n")
-
-    write.append("\nFormat:\n\tstring_size: [runtime1, ..., runtimeN]")
-    footer = "\tNOTE: Runtimes measured in"
-
-    if micro_sec:
-        footer += " microseconds (μs)"
-    else:
-        footer += " nanoseconds (ns)"
-
-    write.append(footer)
 
     return '\n'.join(write)
